+++ plugins/builds/create.js
@@ -24,6 +24,8 @@ module.exports = () => ({
             const jobFactory = request.server.app.jobFactory;
             const buildFactory = request.server.app.buildFactory;
             const userFactory = request.server.app.userFactory;
+            const eventFactory = request.server.app.eventFactory;
+            const scm = buildFactory.scm;
             const username = request.auth.credentials.username;
             const payload = {
                 jobId: request.payload.jobId,
@@ -40,16 +42,29 @@ module.exports = () => ({
                 job.pipeline.then(pipeline =>
                     // ask the user for permissions on this repo
                     user.getPermissions(pipeline.scmUri)
-                        // check if user has push access
-                        .then((permissions) => {
-                            if (!permissions.push) {
-                                throw boom.unauthorized(`User ${username} `
-                                    + 'does not have push permission for this repo');
-                            }
-                        })
-                        // user has good permissions, sync and create a build
-                        .then(() => pipeline.sync())
-                        .then(() => buildFactory.create(payload))
+                    // check if user has push access
+                    .then((permissions) => {
+                        if (!permissions.push) {
+                            throw boom.unauthorized(`User ${username} `
+                                + 'does not have push permission for this repo');
+                        }
+                    })
+                    // user has good permissions, sync and create a build
+                    .then(() => pipeline.sync())
+                    .then(() => user.unsealToken())
+                    .then(token => scm.getCommitSha({ token, scmUri: pipeline.scmUri }))
+                    .then(sha => eventFactory.create({
+                        pipelineId: pipeline.id,
+                        type: job.isPR() ? 'pr' : 'pipeline',
+                        workflow: job.isPR() ? ['main'] : pipeline.workflow,
+                        username,
+                        sha
+                    }))
+                    .then((event) => {
+                        payload.eventId = event.id;
+
+                        return buildFactory.create(payload);
+                    })
             )).then((build) => {
                 // everything succeeded, inform the user
                 const location = urlLib.format({